+++ Cocktails.Data.EFCore/Migrations/Identity/IdentityContextModelSnapshot.cs
@@ -35,7 +35,7 @@ namespace Cocktails.Data.EFCore.Migrations.Identity
 
             modelBuilder.Entity("Cocktails.Data.Domain.RoleClaim", b =>
                 {
-                    b.Property<int>("Id")
+                    b.Property<long>("Id")
                         .ValueGeneratedOnAdd();
 
                     b.Property<string>("ClaimType");
@@ -73,7 +73,7 @@ namespace Cocktails.Data.EFCore.Migrations.Identity
 
             modelBuilder.Entity("Cocktails.Data.Domain.UserClaim", b =>
                 {
-                    b.Property<int>("Id")
+                    b.Property<long>("Id")
                         .ValueGeneratedOnAdd();
 
                     b.Property<string>("ClaimType");
